python/agents/RAG/rag_tools_package/pyproject.toml
[project]
name = "rag-tools"
version = "0.1.0"
description = "Tools for RAG ingestion and CPT extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "google-cloud-aiplatform",
    "EbookLib",
    "beautifulsoup4",
    "python-dotenv",
    "requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_tools"]
